:bug: Patch isCursorOnInterpolatedString
The problem lies in the fact that "scopes at cursor" is actually the
scopes of the character on the right, so to determine if the cursor is
on an interpolated string, the character on the other side of the cursor
should also be checked, with two exceptions: when the cursor is at the
beginning or end of a line.

This patch aims to be clean, but not perfect:

** In some rare situations, such as when typing `#` at the beginning or
end of a line within a multi-line interpolated string, it won't expand
to `#{}`. **