Fix tag-finder for self-closing tags with '>' in their attributes
Also:
- Fix a bug in an existing spec concerning the matching of self
  closing tags (wrong cursor position)
- Add specs for self-closing tags after the cursor

The problem that is addressed here is that `<foo bar=">"/>` was
previously catched as a normal opening tag. Therfore `(?![^>]*\/>)`
was replaced with `(?:\"[^\"]*\"|'[^']*'|[^>\\/\"'])*>` and to remove
redundency and coplexity the function generateTagStartOrEndRegex was
introduced to generate the affected regexes.

Because we (@ruv and I discussed the regex in the PR) decided to stop
using a negative lookahead in favor of performance and simplicity,
this changed the behavior of the function tagsNotClosedInFragment when
it encounters malformed XML and broke an existing spec that verified
this. However because this behavior was already criticized as not
desirable, I decided to remove the spec.